+++ src/definitions/model.ts
@@ -2,7 +2,6 @@ import { CliCommand, type ParentCliCommandDefinition } from '@causa/cli';
 import { WorkspaceFunction } from '@causa/workspace';
 import { AllowMissing } from '@causa/workspace/validation';
 import { Allow, IsArray, IsObject, IsString } from 'class-validator';
-import type { InputData } from 'quicktype-core';
 import type {
   LoadSchemasResult,
   ObjectSchemaWithoutDatabases,
@@ -147,21 +146,6 @@ export type CodeGeneratorInputs = {
  */
 export abstract class ModelParseCodeGeneratorInputs extends WorkspaceFunction<
   Promise<CodeGeneratorInputs>
-> {
-  /**
-   * The configuration for the code generator.
-   */
-  @IsObject()
-  readonly configuration!: Record<string, unknown>;
-}
-
-/**
- * Creates `quicktype`'s {@link InputData} for model code generators, based on their configuration.
- * This uses the {@link ModelParseCodeGeneratorInputs} to get the list of files and options, but the quicktype's input
- * data will depend on the model's schema format.
- */
-export abstract class ModelMakeGeneratorQuicktypeInputData extends WorkspaceFunction<
-  Promise<InputData>
 > {
   /**
    * The configuration for the code generator.